Feat: Implemented LI, fixed SLL
Load immediate generates a res_zeroextimm signal that causes the ZeroExtImm to  be written directly to the regfile, bypassing the ALU. The other way this can be done is to add a zero-extender and create another MUX for ALUsrca and ALUsrcb that zeroes srca and gives the ZeroExtImm to srcb.

Also fixed SLL to shift srcb, not srca, by shamt.

test_sb.mem contains the following instructions:
```
li      $1, 0xbbaa
li      $2, 0xb0b0
sll     $1, $1, 16
or      $3, $1, $2
addi    $5, $0, 16
sw      $3, 0($5)
stop
```